Flyttet eierskap fra modal til dropdown.
